ISS90 - commit auc computation
